wip: stabilizing event exception data.
